Made Paperless-ngx Script
Downloads and installs Paperless-ngx including JBIG2.
Makes a paperless.creds file in root home directory that contains the usernames and passwords for Postgres Database and the WebUI.
Tested on Debian 11, Ubuntu 20.04,21.10,22.04
Debian 10 has too many errors and qwerks to be usable.
Debian 11 needs nesting and keyctl if not privileged.
Ubuntu doesn't need either if not privileged.

redis doesn't start if you remove the {cache,log} directories, so I removed that line.