Update monitor-all.sh
Bugs have been fixed, e.g. the IP address was taken over from the previous VM if none was found.
The parameters for "onboot" were not recognized correctly.
Furthermore, the IP search string was packed into a variable.
The output has been extended by the name of the PCT/VM.
A maximum of 3 restarts are preset, but this can also be adjusted using a variable. The original always restarts.
Emails can now also be sent for restarts.