fix: install Debian 12 version of MS deps, add libmsquic
The script was installing the Debian 11 version of the Microsoft packages repo on Debian 12, which isn't officially supported. Additionally, installing `libmsquic` allows Technitium DNS to serve DNS-over-QUIC, but this fails if the Debian 11 repo is used on Debian 12 (because it wants to install libssl-1.1, which is EOL and removed from Debian 12).

This PR fixes this issue by changing the version of the MS repo that's installed from 11 to 12. Additionally, this installs `libmsquic`, because you really might as well have that feature. Lastly, I condensed the `apt-get install` commands into one command after installing the MS repo, which will make the script run more quickly. 

Thanks for the useful scripts!